Package qemu
On macOS, this builds qemu with homebrew, based on upstream formula but
limited to just what we need.  This is so we can reuse finch's script to
identify the required files.

On Linux, we build qemu directly, and use the scripts from the old
lima-and-qemu repo to identify the files we need to ship.

Because the aarch64 macOS runners on GitHub do not support
virtualization, we hack in a few extra files to include on that
platform.  The list of files is based on what the x86_64 run had.

We need slirp because lima uses that (with `--netdev user,...`).

+++ appdir-qemu.sh
@@ -1,8 +1,24 @@
 #!/bin/bash
 
-# This is a helper script to build an appDir structure to
-# include lima and qemu as part of an AppImage binary.
-# The binaries build should happen in the oldest available Ubuntu LTS
+# Copyright 2024 SUSE LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+# From https://github.com/rancher-sandbox/lima-and-qemu/blob/d874fe62584dcfa7e5fdd45f5225853a0240bd26/bin/appdir-lima-and-qemu.sh
+
+# This is a helper script to build an appDir structure to include qemu as part
+# of an AppImage binary.  The binaries build should happen in the oldest
+# available Ubuntu LTS.
 
 function error {
   >&2 echo "$@"
@@ -14,22 +30,23 @@ function keepListedFiles {
   local list=$2
 
   [ -d "${dir}" ] || error "${dir} is not a directory"
-  
-  for it in $(ls ${dir}); do
+
+  for it in "${dir}"/*; do
+    it=${it#"$dir"/}
     if [[ "${list}" =~ \ ${it}\  ]]; then
       continue
     fi
-    [ -n "${it}" ] && rm -rf "${dir}/${it}"
+    [ -n "${it}" ] && rm -rf "${dir:?}/${it}"
   done
 }
 
-set -e
+set -ex
 
 [ -n "${1}" ] || error "One argument to the built app dir is required"
 [ -d "${1}" ] || error "Directory ${1} doesn't exist"
 
 appDir=$1
-dist="lima-and-qemu"
+dist="qemu-linux-x86_64"
 
 # Inspired on linuxdeployqt https://github.com/probonopd/linuxdeployqt/blob/master/tools/linuxdeployqt/excludelist.h
 # Linuxdeployqt is a tool created by probonopd, the AppImage creator
@@ -50,7 +67,7 @@ excludeLibs+="libdl.so.2 "
 excludeLibs+="libgmodule-2.0.so.0 "
 
 firmwareOfInterest=" bios-256k.bin edk2-x86_64-code.fd efi-virtio.rom kvmvapic.bin vgabios-virtio.bin "
-executablesOfInterest=" qemu-system-x86_64 qemu-img limactl "
+executablesOfInterest=" qemu-system-x86_64 qemu-img "
 
 mkdir -p "${appDir}/lib"
 
@@ -64,18 +81,15 @@ for lib in $(echo "${linkedLibs}" | sort | uniq ); do
 done
 
 # strip docs
-rm -rf "${appDir}/share/doc"
-
-# remove lima agent for aarch64
-rm "${appDir}/share/lima/lima-guestagent.Linux-aarch64" 
+rm -rf "${appDir:?}/share/doc"
 
 # remove qemu icons, includes, etc.
 # We could fine tune the firmaware files
-rm -rf "${appDir}/include"
-rm -rf "${appDir}/libexec"
-rm -rf "${appDir}/var"
-rm -rf "${appDir}/share/applications"
-rm -rf "${appDir}/share/icons"
+rm -rf "${appDir:?}/include"
+rm -rf "${appDir:?}/libexec"
+rm -rf "${appDir:?}/var"
+rm -rf "${appDir:?}/share/applications"
+rm -rf "${appDir:?}/share/icons"
 
 # keep only relevant firmware
 keepListedFiles "${appDir}/share/qemu" "${firmwareOfInterest}"
